refactor: disallow guest mode only for embedded page
